requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==2.0.4
langchain-ollama==0.2.1
chromadb==0.5.18
pypdf==4.3.1
python-dotenv==1.0.1
pydantic==2.9.2
